+++ src/test/java/com/chroma/pages/RealEstateAdminIndexPage.java
@@ -3,7 +3,6 @@ package com.chroma.pages;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
-
 import com.chroma.web.WebDriverUtils;
 
 public class RealEstateAdminIndexPage {
