+++ src/test/java/com/chroma/runners/CTSMSSmokeRunners.java
@@ -1,7 +1,6 @@
 package com.chroma.runners;
 
 import org.junit.runner.RunWith;
-
 import cucumber.api.CucumberOptions;
 import cucumber.api.junit.Cucumber;
 
@@ -12,7 +11,7 @@ import cucumber.api.junit.Cucumber;
 		features = "src/test/java/com/chroma/features", 
 		glue = "com.chroma.stepDefinitions", 
 		tags = "@CTSMSSmoke", 
-		dryRun =  false, 
+		dryRun = false, 
 		monochrome = true, 
 		strict = true
 )
