+++ src/test/java/com/chroma/runners/RealEstateProgressionRunners.java
@@ -1,7 +1,6 @@
 package com.chroma.runners;
 
 import org.junit.runner.RunWith;
-
 import cucumber.api.CucumberOptions;
 import cucumber.api.junit.Cucumber;
