+++ src/test/java/com/chroma/runners/RealEstateRegressionRunners.java
@@ -1,7 +1,6 @@
 package com.chroma.runners;
 
 import org.junit.runner.RunWith;
-
 import cucumber.api.CucumberOptions;
 import cucumber.api.junit.Cucumber;
 
@@ -12,11 +11,11 @@ import cucumber.api.junit.Cucumber;
 		features = "src/test/java/com/chroma/features", 
 		glue = "com.chroma.stepDefinitions", 
 		tags = "@RealEstateRegression", 
-		dryRun =  false, 
+		dryRun = false, 
 		monochrome = true, 
 		strict = true
 ) 
 
-public class RealEstateRefressionRunners {
+public class RealEstateRegressionRunners {
     
 }
