+++ src/test/java/com/chroma/runners/RealEstateSmoke.java
@@ -1,7 +1,6 @@
 package com.chroma.runners;
 
 import org.junit.runner.RunWith;
-
 import cucumber.api.CucumberOptions;
 import cucumber.api.junit.Cucumber;
 
@@ -11,8 +10,8 @@ import cucumber.api.junit.Cucumber;
 		"pretty", "com.aventstack.extentreports.cucumber.adapter.ExtentCucumberAdapter:"  }, 
 		features = "src/test/java/com/chroma/features", 
 		glue = "com.chroma.stepDefinitions", 
-		tags = "@123", 
-		dryRun =  false, 
+		tags = "@Smoke", 
+		dryRun = false, 
 		monochrome = true, 
 		strict = true
 )
