+++ src/test/java/com/chroma/stepDefinitions/StudentAdmissionStepDef.java
@@ -1,7 +1,6 @@
 package com.chroma.stepDefinitions;
 
 import org.openqa.selenium.JavascriptExecutor;
-
 import com.chroma.appsCommon.PageInitializer;
 import com.chroma.utils.CucumberLogUtils;
 import com.chroma.web.CommonUtils;
